+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/FeatureFlagsTask.kt
@@ -1,5 +1,6 @@
 package io.mehow.laboratory.gradle
 
+import io.mehow.laboratory.generator.FeatureFlagModel
 import javax.inject.Inject
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.model.ObjectFactory
@@ -20,8 +21,23 @@ internal abstract class FeatureFlagsTask @Inject constructor(objects: ObjectFact
   fun generateFeatureFlags() {
     outputDirectory.get().asFile.deleteRecursively()
     val outputPath = outputDirectory.get().asFile
-    inputFlags.get().map(FeatureFlagInput::toModel).forEach { model ->
-      model.prepare().writeTo(outputPath)
+    requireNoDuplicates().forEach { model -> model.prepare().writeTo(outputPath) }
+  }
+
+  private fun requireNoDuplicates(): List<FeatureFlagModel> {
+    val models = inputFlags.get().map(FeatureFlagInput::toModel)
+    val groupedModels = models.groupBy(FeatureFlagModel::className)
+    require(models.size == groupedModels.size) {
+      val duplicates =
+        groupedModels
+          .filterValues { it.size > 1 }
+          .mapValues { (_, models) -> models.map { it.className.canonicalName } }
+          .values
+          .flatten()
+          .distinct()
+          .sorted()
+      "Feature flags must have unique fully qualified names. Found following duplicates: $duplicates"
     }
+    return models
   }
 }