+++ laboratory/gradle-plugin/src/test/kotlin/io/mehow/laboratory/gradle/test/BuildResult.kt
@@ -52,6 +52,9 @@ class Fixture(private val projectDir: File, private val expectFailure: Boolean) 
 }
 
 class BuildResult(private val projectDir: File, private val result: GradleBuildResult) {
+  val output
+    get() = result.output
+
   operator fun get(task: LaboratoryTask): TaskOutcome {
     val task =
       withClue("Expected '${projectDir.name}' project to have ${task.taskName} task") {